requirements.txt
altair==6.0.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fonttools==4.61.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
imbalanced-learn==0.14.0
imblearn==0.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
narwhals==2.14.0
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.52.2
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
watchdog==6.0.0
wcwidth==0.2.14
